@@ -16,10 +16,10 @@
     "jose": "^6.0.11",
     "js-cookie": "^3.0.5",
     "mysql2": "^3.14.1",
-    "next": "15.3.1",
+    "next": "^16.1.1",
     "nodemailer": "^7.0.11",
-    "react": "^19.0.0",
-    "react-dom": "^19.0.0",
+    "react": "^19.2.3",
+    "react-dom": "^19.2.3",
     "react-hot-toast": "^2.5.2",
     "react-modal": "^3.16.3",
     "react-toastify": "^11.0.5"
